Organized campaign cards into individual components
