+++ ui/components/index.ts
@@ -1,9 +1,12 @@
+export * from "./BalanceRefundCard"
 export * from "./Button"
 export * from "./ButtonLink"
 export * from "./CampaignCards"
 export * from "./CampaignDetails"
 export * from "./CampaignUtils"
 export * from "./CenteredColumn"
+export * from "./FundPendingCard"
+export * from "./GovernanceCard"
 export * from "./Header"
 export * from "./Input"
 export * from "./Loader"
@@ -13,4 +16,3 @@ export * from "./StatusIndicator"
 export * from "./Suspense"
 export * from "./TooltipInfo"
 export * from "./WalletMessage"
-export * from "./PendingCard"
