+++ ui/pages/campaign/[address].tsx
@@ -1,14 +1,11 @@
-import { ArcElement, Chart as ChartJS } from "chart.js"
 import cn from "classnames"
 import type { NextPage } from "next"
 import { NextRouter, useRouter } from "next/router"
-import { FC, useCallback, useEffect, useState } from "react"
-import { Pie } from "react-chartjs-2"
-import { useForm } from "react-hook-form"
-import { useRecoilValue, useSetRecoilState } from "recoil"
+import { FC, useEffect, useState } from "react"
+import { useRecoilValue } from "recoil"
 
 import {
-  Button,
+  BalanceRefundCard,
   ButtonLink,
   CampaignAction,
   CampaignDetails,
@@ -16,89 +13,20 @@ import {
   CampaignProgress,
   CampaignStatus,
   CenteredColumn,
-  ControlledFormPercentTokenDoubleInput,
-  FormInput,
+  FundPendingCard,
+  GovernanceCard,
   Loader,
   ResponsiveDecoration,
   Suspense,
   WalletMessage,
-  PendingCard,
 } from "../../components"
-import { daoUrlPrefix, payTokenSymbol, theme } from "../../helpers/config"
-import { numberPattern } from "../../helpers/form"
+import { ContributeCard } from "../../components/ContributeCard"
+import { payTokenSymbol, theme } from "../../helpers/config"
 import { prettyPrintDecimal } from "../../helpers/number"
-import { useContributeCampaign } from "../../hooks/useContributeCampaign"
-import { useCopy } from "../../hooks/useCopy"
-import { useFundPendingCampaign } from "../../hooks/useFundPendingCampaign"
-import { useRefundCampaign } from "../../hooks/useRefundCampaign"
 import { useWallet } from "../../hooks/useWallet"
 import { suggestToken } from "../../services/keplr"
-import {
-  campaignWalletBalance,
-  favoriteCampaignAddressesAtom,
-  fetchCampaign,
-  fetchCampaignActions,
-} from "../../state/campaigns"
+import { fetchCampaign, fetchCampaignActions } from "../../state/campaigns"
 import { Status } from "../../types"
-
-ChartJS.register(ArcElement)
-
-interface AddressDisplayProps {
-  label: string
-  address: string
-}
-
-const AddressDisplay: FC<AddressDisplayProps> = ({ label, address }) => {
-  const { copy, Icon } = useCopy(address)
-
-  return (
-    <p
-      className="mt-1 first-of-type:mt-4 flex flex-row justify-between items-center hover:opacity-70 transition cursor-pointer select-none"
-      onClick={copy}
-    >
-      <span className="text-placeholder pr-2">{label}</span>
-      <span className="text-green opacity-70 font-mono flex flex-row items-center">
-        <Icon size={28} className="pr-2" />
-        {address.substring(0, 8)}
-        ...
-        {address.substring(address.length - 6, address.length)}
-      </span>
-    </p>
-  )
-}
-
-interface PieLegendProps {
-  items: {
-    label: string
-    color: string
-  }[]
-  className?: string
-}
-const PieLegend: FC<PieLegendProps> = ({ items, className }) => (
-  <div className={cn("flex flex-col", className)}>
-    {items.map(({ label, color: backgroundColor }) => (
-      <div key={label} className="flex flex-row items-center mt-1 first:mt-0">
-        <div
-          className={cn("w-10 h-5 mr-2 shrink-0")}
-          style={{ backgroundColor }}
-        ></div>
-        <p className="text-light">{label}</p>
-      </div>
-    ))}
-  </div>
-)
-
-interface FundPendingForm {
-  tokens?: number
-}
-
-interface ContributionForm {
-  contribution?: number
-}
-
-interface RefundForm {
-  refund?: number
-}
 
 export const Campaign: NextPage = () => {
   const router = useRouter()
@@ -133,9 +61,11 @@ interface CampaignContentProps {
 const CampaignContent: FC<CampaignContentProps> = ({
   router: { isReady, query, push: routerPush },
 }) => {
-  const { connected, keplr } = useWallet()
   const campaignAddress =
     isReady && typeof query.address === "string" ? query.address : ""
+
+  const { keplr, connected } = useWallet()
+
   const { campaign, error: campaignError } = useRecoilValue(
     fetchCampaign(campaignAddress)
   )
@@ -143,50 +73,6 @@ const CampaignContent: FC<CampaignContentProps> = ({
   const { actions, error: campaignActionsError } = useRecoilValue(
     fetchCampaignActions(campaignAddress)
   )
-
-  const { balance: fundingTokenBalance, error: fundingTokenBalanceError } =
-    useRecoilValue(campaignWalletBalance(campaign?.address))
-
-  // Funding form for pending campaigns
-  const {
-    handleSubmit: fundPendingHandleSubmit,
-    register: fundPendingRegister,
-    formState: { errors: fundPendingErrors },
-    watch: fundPendingWatch,
-  } = useForm({ mode: "onChange", defaultValues: {} as FundPendingForm })
-
-  const { fundPendingCampaign, fundPendingCampaignError } =
-    useFundPendingCampaign(campaign)
-  const [fundCampaignProposalUrl, setFundCampaignProposalUrl] = useState(
-    null as string | null
-  )
-
-  // Contribution Form
-  const {
-    handleSubmit: contributionHandleSubmit,
-    register: contributionRegister,
-    formState: { errors: contributionErrors },
-    watch: contributionWatch,
-    reset: contributionReset,
-  } = useForm({
-    defaultValues: {} as ContributionForm,
-  })
-
-  const { contributeCampaign, contributeCampaignError } =
-    useContributeCampaign(campaign)
-
-  // Refund Form
-  const {
-    handleSubmit: refundHandleSubmit,
-    formState: { errors: refundErrors },
-    watch: refundWatch,
-    control: refundControl,
-    reset: refundReset,
-  } = useForm({
-    defaultValues: {} as RefundForm,
-  })
-
-  const { refundCampaign, refundCampaignError } = useRefundCampaign(campaign)
 
   // If no campaign, navigate to campaigns list.
   useEffect(() => {
@@ -197,16 +83,6 @@ const CampaignContent: FC<CampaignContentProps> = ({
   const [showAddFundingToken, setShowAddFundingToken] = useState(false)
   const [showAddGovToken, setShowAddGovToken] = useState(false)
 
-  // Add campaign to favorites if funding.
-  const setFavoriteAddresses = useSetRecoilState(favoriteCampaignAddressesAtom)
-  const addFavorite = useCallback(
-    (address: string) =>
-      setFavoriteAddresses((addresses) =>
-        addresses.includes(address) ? addresses : [...addresses, address]
-      ),
-    [setFavoriteAddresses]
-  )
-
   // If page not ready, display loader.
   if (!isReady) return <Loader overlay />
   // Display nothing (redirecting to campaigns list, so this is just a type check).
@@ -214,25 +90,14 @@ const CampaignContent: FC<CampaignContentProps> = ({
 
   const {
     name,
-
     status,
 
     goal,
     pledged,
+
     dao: {
       url: daoUrl,
-      govToken: {
-        symbol: govTokenSymbol,
-        campaignBalance: govTokenCampaignBalance,
-        daoBalance: govTokenDAOTreasuryBalance,
-        supply: govTokenSupply,
-      },
-    },
-
-    fundingToken: {
-      symbol: tokenSymbol,
-      price: fundingTokenPrice,
-      supply: fundingTokenSupply,
+      govToken: { address: govTokenAddress },
     },
   } = campaign ?? {}
 
@@ -242,87 +107,7 @@ const CampaignContent: FC<CampaignContentProps> = ({
       !(await suggestToken(keplr, campaign.fundingToken.address))
     )
   const suggestGovToken = async () =>
-    keplr &&
-    setShowAddGovToken(
-      !(await suggestToken(keplr, campaign.dao.govToken.address))
-    )
-
-  // Contribution Form
-  const watchContribution = contributionWatch("contribution")
-  const doContribution = async ({ contribution }: ContributionForm) => {
-    if (!contribution) return
-
-    // TODO: Add success display.
-    if (await contributeCampaign(contribution)) {
-      // Attempt to add token to Keplr.
-      await suggestFundingToken()
-
-      // Add campaign to favorites.
-      addFavorite(campaignAddress)
-
-      // Empty form fields.
-      contributionReset()
-    }
-  }
-
-  // Refund Form
-  const watchRefund = refundWatch("refund")
-  const doRefund = async ({ refund }: RefundForm) => {
-    // If funded, the refund action becomes the join DAO action, so send all tokens.
-    if (status === Status.Funded) refund = fundingTokenBalance ?? 0
-
-    if (!refund) return
-
-    // TODO: Add success display.
-    if (await refundCampaign(refund)) {
-      // Attempt to add token to Keplr if receiving governance tokens.
-      if (status === Status.Funded) await suggestGovToken()
-
-      // Empty form fields.
-      refundReset()
-    }
-  }
-
-  // DAO voting power of campaign (determined by proportion of campaign's governance token balance to all governance tokens not in the DAO's treasury).
-  const campaignVotingPower =
-    govTokenCampaignBalance &&
-    govTokenSupply &&
-    govTokenDAOTreasuryBalance &&
-    govTokenSupply > govTokenDAOTreasuryBalance
-      ? (100 * govTokenCampaignBalance) /
-        (govTokenSupply - govTokenDAOTreasuryBalance)
-      : undefined
-  // DAO voting power of existing DAO members.
-  const govTokenDAOMemberBalance =
-    govTokenSupply &&
-    govTokenDAOTreasuryBalance !== undefined &&
-    govTokenCampaignBalance
-      ? govTokenSupply - govTokenDAOTreasuryBalance - govTokenCampaignBalance
-      : undefined
-
-  // Percent of funding tokens the user's balance represents.
-  const fundingTokenBalancePercent =
-    fundingTokenBalance && fundingTokenSupply
-      ? (100 * fundingTokenBalance) / fundingTokenSupply
-      : undefined
-
-  // Contribution
-  const expectedFundingTokensReceived =
-    watchContribution && watchContribution > 0 && fundingTokenPrice
-      ? fundingTokenPrice * watchContribution
-      : 0
-  // Max contribution is remaining amount left to fund. Cannot fund more than goal.
-  const maxContribution = Math.min(
-    goal - pledged,
-    Number.MAX_SAFE_INTEGER / 1e6
-  )
-  // Refund
-  // Minimum refund is how many funding tokens (with decimals) per 1 ujuno(x).
-  const minRefund = Math.ceil(fundingTokenPrice ?? 1e-6) / 1e6
-  const expectedPayTokensReceived =
-    watchRefund && watchRefund > 0 && fundingTokenPrice
-      ? watchRefund / fundingTokenPrice
-      : 0
+    keplr && setShowAddGovToken(!(await suggestToken(keplr, govTokenAddress)))
 
   return (
     <>
@@ -354,60 +139,15 @@ const CampaignContent: FC<CampaignContentProps> = ({
           >
             <CampaignDetails {...campaign} />
 
-            {connected || <WalletMessage />}
+            {!connected && <WalletMessage />}
 
             {status === Status.Pending ? (
-              <PendingCard campaign={campaign} />
+              <FundPendingCard campaign={campaign} />
             ) : status === Status.Open ? (
-              <form
-                onSubmit={contributionHandleSubmit(doContribution)}
-                className={cn(
-                  "flex flex-col items-stretch mt-8",
-                  "sm:flex-row sm:items-start lg:self-stretch lg:mb-0"
-                )}
-              >
-                <FormInput
-                  type="number"
-                  inputMode="decimal"
-                  placeholder="Contribute..."
-                  accent={
-                    expectedFundingTokensReceived
-                      ? `You will receive about ${prettyPrintDecimal(
-                          expectedFundingTokensReceived
-                        )} ${tokenSymbol}`
-                      : undefined
-                  }
-                  wrapperClassName="!mb-4 sm:!mb-0 sm:mr-4 sm:flex-1"
-                  className="!py-3 !px-6 !pr-28"
-                  tail={payTokenSymbol}
-                  error={
-                    contributionErrors?.contribution?.message ??
-                    contributeCampaignError ??
-                    undefined
-                  }
-                  disabled={!connected}
-                  {...contributionRegister("contribution", {
-                    valueAsNumber: true,
-                    pattern: numberPattern,
-                    min: {
-                      value: 1e-6,
-                      message: `Must be at least 0.000001 ${payTokenSymbol}.`,
-                    },
-                    max: {
-                      value: maxContribution,
-                      message: `Must be less than or equal to ${prettyPrintDecimal(
-                        maxContribution
-                      )} ${payTokenSymbol}.`,
-                    },
-                  })}
-                />
-
-                <Button
-                  className="sm:h-[50px]"
-                  disabled={!connected}
-                  submitLabel="Support this campaign"
-                />
-              </form>
+              <ContributeCard
+                campaign={campaign}
+                suggestFundingToken={suggestFundingToken}
+              />
             ) : undefined}
           </div>
           <div className="flex flex-col flex-wrap self-stretch">
@@ -449,197 +189,19 @@ const CampaignContent: FC<CampaignContentProps> = ({
                 <p className="text-light text-sm">Backers</p> */}
             </div>
 
-            {/* Hide for funded campaigns since govTokenCampaignBalance won't remain constant. */}
-            {/* TODO: Store initial fund amount in contract staet and use that instead. */}
-            {status !== Status.Funded &&
-              !!campaignVotingPower &&
-              !!govTokenDAOTreasuryBalance &&
-              !!govTokenCampaignBalance &&
-              !!govTokenSupply &&
-              !!govTokenSymbol && (
-                <div
-                  className={cn(
-                    "bg-card rounded-3xl p-8 mt-4 lg:mt-8",
-                    "flex flex-col items-start",
-                    "max-w-full"
-                  )}
-                >
-                  <h3 className="text-green text-3xl">
-                    {prettyPrintDecimal(campaignVotingPower, 2)}% governance
-                  </h3>
-                  <p className="text-light text-sm">
-                    Campaign backers will have{" "}
-                    {prettyPrintDecimal(campaignVotingPower, 2)}% voting power
-                    in the DAO. Voting power ignores the DAO&apos;s treasury
-                    balance. To learn more,{" "}
-                    <a
-                      href="https://docs.daoup.zone/evaluating-campaigns#what-is-a-good-percentage-of-governance-power"
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      className="underline hover:no-underline"
-                    >
-                      read the docs
-                    </a>
-                    .
-                  </p>
-
-                  <Pie
-                    options={{
-                      // Disable all events (hover, tooltip, etc.)
-                      events: [],
-                      animation: false,
-                    }}
-                    data={{
-                      datasets: [
-                        {
-                          data: [
-                            govTokenDAOMemberBalance,
-                            govTokenDAOTreasuryBalance,
-                            govTokenCampaignBalance,
-                          ],
-                          backgroundColor: [
-                            theme.colors.pieMedium,
-                            theme.colors.pieDark,
-                            theme.colors.pieLight,
-                          ],
-                          borderWidth: 0,
-                        },
-                      ],
-                    }}
-                    className="!w-48 !h-48 mt-8 self-center"
-                  />
-
-                  <PieLegend
-                    className="mt-4"
-                    items={[
-                      {
-                        label: "Campaign",
-                        color: theme.colors.pieLight,
-                      },
-                      {
-                        label: "Current DAO members/creators",
-                        color: theme.colors.pieMedium,
-                      },
-                      {
-                        label: "DAO's treasury",
-                        color: theme.colors.pieDark,
-                      },
-                    ]}
-                  />
-                  <div className="flex flex-col"></div>
-                </div>
-              )}
+            {/* TODO: Show for funded campaigns by storing initial fund amount in contract state and use that instead (since govTokenCampaignBalance won't remain constant). */}
+            {status === Status.Open && <GovernanceCard campaign={campaign} />}
           </div>
         </div>
 
         {connected && (
-          <div
-            className={cn(
-              "bg-card rounded-3xl",
-              "mt-4 lg:mt-8",
-              "w-full lg:w-3/5",
-              "py-8 px-12"
-            )}
-          >
-            <h2 className="text-xl text-green mb-2">Your Balance</h2>
-
-            <p className="text-light">
-              {prettyPrintDecimal(fundingTokenBalance ?? 0)} {tokenSymbol}
-              {fundingTokenBalancePercent && (
-                <span className="text-placeholder ml-2">
-                  {prettyPrintDecimal(fundingTokenBalancePercent, 2)}% of total
-                  supply
-                </span>
-              )}
-            </p>
-            {fundingTokenBalance !== null &&
-              fundingTokenBalance > 0 &&
-              showAddFundingToken && (
-                <div className="mt-4">
-                  <Button onClick={suggestFundingToken}>
-                    Add token to wallet
-                  </Button>
-                  <p className="text-sm text-placeholder italic mt-2">
-                    This allows you to view your campaign funding token balance
-                    ({tokenSymbol}) from your Keplr wallet. If you&apos;ve
-                    already done this, it should still be there.
-                  </p>
-                </div>
-              )}
-            {status === Status.Funded && showAddGovToken && (
-              <div className="mt-4">
-                <Button onClick={suggestGovToken}>
-                  Add DAO token to wallet
-                </Button>
-                <p className="text-sm text-placeholder italic mt-2">
-                  This allows you to view your DAO governance token balance from
-                  your Keplr wallet. If you&apos;ve already done this, it should
-                  still be there.
-                </p>
-              </div>
-            )}
-
-            {status !== Status.Pending &&
-              fundingTokenBalance !== null &&
-              fundingTokenBalance > 0 && (
-                <>
-                  {status !== Status.Funded && (
-                    <h2 className="text-xl text-green mt-8 mb-4">Refunds</h2>
-                  )}
-
-                  <form onSubmit={refundHandleSubmit(doRefund)}>
-                    {status === Status.Funded ? (
-                      <p className="mt-4 text-placeholder italic">
-                        This campaign has been successfully funded. To join the
-                        DAO, exchange your {tokenSymbol} tokens by clicking the
-                        button below.
-                      </p>
-                    ) : (
-                      <ControlledFormPercentTokenDoubleInput
-                        name="refund"
-                        control={refundControl}
-                        minValue={minRefund}
-                        maxValue={fundingTokenBalance}
-                        currency={tokenSymbol}
-                        first={{
-                          placeholder: "50",
-                        }}
-                        second={{
-                          placeholder: prettyPrintDecimal(
-                            fundingTokenBalance * 0.5
-                          ),
-                        }}
-                        shared={{
-                          disabled: status !== Status.Open,
-                        }}
-                        accent={
-                          expectedPayTokensReceived
-                            ? `You will receive about ${prettyPrintDecimal(
-                                expectedPayTokensReceived
-                              )} ${payTokenSymbol}`
-                            : undefined
-                        }
-                        error={
-                          refundErrors?.refund?.message ??
-                          refundCampaignError ??
-                          undefined
-                        }
-                      />
-                    )}
-
-                    <Button
-                      submitLabel={
-                        status === Status.Funded ? "Join DAO" : "Refund"
-                      }
-                      className="mt-4"
-                      disabled={
-                        status !== Status.Open && status !== Status.Funded
-                      }
-                    />
-                  </form>
-                </>
-              )}
-          </div>
+          <BalanceRefundCard
+            campaign={campaign}
+            showAddGovToken={showAddGovToken}
+            suggestGovToken={suggestGovToken}
+            showAddFundingToken={showAddFundingToken}
+            suggestFundingToken={suggestFundingToken}
+          />
         )}
 
         <h2 className="text-green text-xl mt-8">Activity</h2>
